+++ src/main/java/com/github/razorplay01/cpa/ModTemplate.java
@@ -22,21 +22,21 @@ import com.github.razorplay01.cpa.platform.fabric.FabricPlatform;
 *///?}
 
 //? if <= 1.21.10 {
-import net.minecraft.resources.ResourceLocation;
-//?}
+/*import net.minecraft.resources.ResourceLocation;
+*///?}
 
 @SuppressWarnings("LoggingSimilarMessage")
 public class ModTemplate {
 	public static final String MOD_ID = /*$ mod_id*/ "cpa";
-	public static final String MOD_VERSION = /*$ mod_version*/ "5.2.1";
+	public static final String MOD_VERSION = /*$ mod_version*/ "5.3.0";
 	public static final String MOD_FRIENDLY_NAME = /*$ mod_name*/ "Custom Player Animations";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
 	public static ClientConfig CONFIG;
 
-	public static final /*? if <=1.21.10 {*/ResourceLocation/*?} else {*/ /*net.minecraft.resources.Identifier*//*?}*/ MAIN_ANIMATION_CONTAINER_LAYER_ID = of("main_animation_container");
-	public static final /*? if <=1.21.10 {*/ResourceLocation/*?} else {*/ /*net.minecraft.resources.Identifier*//*?}*/ OVERLAY_ANIMATION_CONTAINER_LAYER_ID = of("overlay_animation_container");
-	public static final /*? if <=1.21.10 {*/ResourceLocation/*?} else {*/ /*net.minecraft.resources.Identifier*//*?}*/ SPECIAL_ANIMATION_CONTAINER_LAYER_ID = of("special_animation_container");
+	public static final /*? if <=1.21.10 {*//*ResourceLocation*//*?} else {*/ net.minecraft.resources.Identifier/*?}*/ MAIN_ANIMATION_CONTAINER_LAYER_ID = of("main_animation_container");
+	public static final /*? if <=1.21.10 {*//*ResourceLocation*//*?} else {*/ net.minecraft.resources.Identifier/*?}*/ OVERLAY_ANIMATION_CONTAINER_LAYER_ID = of("overlay_animation_container");
+	public static final /*? if <=1.21.10 {*//*ResourceLocation*//*?} else {*/ net.minecraft.resources.Identifier/*?}*/ SPECIAL_ANIMATION_CONTAINER_LAYER_ID = of("special_animation_container");
 
 	private static final Platform PLATFORM = createPlatformInstance();
 
@@ -84,7 +84,7 @@ public class ModTemplate {
 		return PlayerAnimResources.getAnimation(of(animationId));
 	}
 
-	public static /*? if <=1.21.10 {*/ResourceLocation/*?} else {*/ /*net.minecraft.resources.Identifier*//*?}*/ of(String path) {
-		return /*? if <=1.21.10 {*/ResourceLocation/*?} else {*/ /*net.minecraft.resources.Identifier*//*?}*/.fromNamespaceAndPath(MOD_ID, path);
+	public static /*? if <=1.21.10 {*//*ResourceLocation*//*?} else {*/ net.minecraft.resources.Identifier/*?}*/ of(String path) {
+		return /*? if <=1.21.10 {*//*ResourceLocation*//*?} else {*/ net.minecraft.resources.Identifier/*?}*/.fromNamespaceAndPath(MOD_ID, path);
 	}
 }