+++ src/main/java/com/github/razorplay01/cpa/mixin/LivingEntityRendererMixin.java
@@ -96,16 +96,16 @@ public abstract class LivingEntityRendererMixin<T extends LivingEntity, S extend
 
 
 			//? if <=1.21.10 {
-			((ExtendedItemStackRenderState) armed.leftHandItem)
+			/*((ExtendedItemStackRenderState) armed.leftHandItem)
 					.setItemStack(livingEntity.getItemHeldByArm(HumanoidArm.LEFT));
 			((ExtendedItemStackRenderState) armed.rightHandItem)
 					.setItemStack(livingEntity.getItemHeldByArm(HumanoidArm.RIGHT));
-			//?} else {
-			/*((ExtendedItemStackRenderState) armed.leftHandItemState)
+			*///?} else {
+			((ExtendedItemStackRenderState) armed.leftHandItemState)
 					.setItemStack(livingEntity.getItemHeldByArm(HumanoidArm.LEFT));
 			((ExtendedItemStackRenderState) armed.rightHandItemState)
 					.setItemStack(livingEntity.getItemHeldByArm(HumanoidArm.RIGHT));
-			*///?}
+			//?}
 		}
 	}
 	//?}