+++ src/main/java/com/github/razorplay01/cpa/platform/common/animation/animations/AnimationProvider.java
@@ -53,6 +53,9 @@ public class AnimationProvider {
 			new CrossbowAnimation(),
 			new GenericHandSwingAnimation(),
 			new SwordAnimation(),
+			//? if >= 1.21.11 {
+			new SpearAnimation(),
+			//?}
 			new PickaxeAnimation(),
 			new AxeAnimation(),
 			new ShovelAnimation()