+++ src/main/java/com/github/razorplay01/cpa/platform/common/animation/animations/base/BoatForwardAnimation.java
@@ -5,10 +5,10 @@ import com.github.razorplay01.cpa.platform.common.util.enums.AnimationsId;
 import com.github.razorplay01.cpa.platform.common.util.interfaces.ICustomAnimation;
 import com.github.razorplay01.cpa.platform.common.util.records.AnimationContext;
 //? if <= 1.21.10 {
-import net.minecraft.world.entity.vehicle.Boat;
-//?}else{
-/*import net.minecraft.world.entity.vehicle.boat.Boat;
-*///?}
+/*import net.minecraft.world.entity.vehicle.Boat;
+*///?}else{
+import net.minecraft.world.entity.vehicle.boat.Boat;
+//?}
 
 import static com.github.razorplay01.cpa.ModTemplate.CONFIG;
 import static com.github.razorplay01.cpa.platform.common.util.Util.configureAnimationContainer;