+++ src/main/java/com/github/razorplay01/cpa/platform/common/util/CustomModifiers.java
@@ -5,7 +5,6 @@ import com.github.razorplay01.cpa.platform.common.util.enums.BodyParts;
 import com.zigythebird.playeranimcore.animation.layered.modifier.AdjustmentModifier;
 import com.zigythebird.playeranimcore.api.firstPerson.FirstPersonMode;
 import com.zigythebird.playeranimcore.math.Vec3f;
-import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.world.entity.HumanoidArm;
 import net.minecraft.world.item.BowItem;
 import net.minecraft.world.item.ShieldItem;
@@ -14,12 +13,18 @@ import java.util.Optional;
 
 import static com.github.razorplay01.cpa.platform.common.util.Util.*;
 
+//? if <= 1.21.8 {
+//import net.minecraft.client.player.AbstractClientPlayer;
+//?} else {
+import net.minecraft.world.entity.Avatar;
+//?}
+
 public class CustomModifiers {
 	private CustomModifiers() {
 		// []
 	}
 
-	public static AdjustmentModifier createBowModifier(/*? if <=1.21.8 {*//*AbstractClientPlayer player*//*?} else {*/ net.minecraft.world.entity.Avatar player/*?}*/) {
+	public static AdjustmentModifier createBowModifier(/*? if <=1.21.8 {*//*AbstractClientPlayer player*//*?} else {*/ Avatar player/*?}*/) {
 		return new AdjustmentModifier(partName -> {
 			boolean isUsingBow = player.isUsingItem() && player.getUseItem().getItem() instanceof BowItem;
 			if (!isUsingBow) return Optional.empty();
@@ -36,7 +41,7 @@ public class CustomModifiers {
 		});
 	}
 
-	public static AdjustmentModifier createShieldModifier(/*? if <=1.21.8 {*//*AbstractClientPlayer player*//*?} else {*/ net.minecraft.world.entity.Avatar player/*?}*/) {
+	public static AdjustmentModifier createShieldModifier(/*? if <=1.21.8 {*//*AbstractClientPlayer player*//*?} else {*/ Avatar player/*?}*/) {
 		return new AdjustmentModifier(partName -> {
 			boolean isUsingShield = player.isUsingItem() && player.getUseItem().getItem() instanceof ShieldItem;
 			if (!isUsingShield) return Optional.empty();
@@ -54,18 +59,18 @@ public class CustomModifiers {
 		});
 	}
 
-	public static AdjustmentModifier createSwingModifier(/*? if <=1.21.8 {*//*AbstractClientPlayer player*//*?} else {*/ net.minecraft.world.entity.Avatar player/*?}*/, AnimationContainer animationContainer) {
+	public static AdjustmentModifier createSwingModifier(/*? if <=1.21.8 {*//*AbstractClientPlayer player*//*?} else {*/ Avatar player/*?}*/, AnimationContainer animationContainer) {
 		return new AdjustmentModifier(partName -> {
 			if (!isSwingingSwordOrTools(player, animationContainer)) {
 				return Optional.empty();
 			}
 
 			float pitchRadians = (float) Math.toRadians(player.getXRot());
-			if (FirstPersonMode.isFirstPersonPass()) {
-				return handleFirstPersonPass(partName, pitchRadians);
-			} else {
-				return handleThirdPersonPass(partName, pitchRadians);
-			}
+			//if (FirstPersonMode.isFirstPersonPass()) {
+			//	return handleFirstPersonPass(partName, pitchRadians);
+			//} else {
+			return handleThirdPersonPass(partName, pitchRadians);
+			//}
 		});
 	}
 }